// src/types/lookup.ts
export interface BusinessCategory {
  businessCategoryId: number;

  categoryName: string;

  isActive: boolean;
}

export interface BusinessType {
  businessTypeId: number;

  businessTypeName: string;
}

export interface Industry {
  industryId: number;

  industryName: string;

  isActive: boolean;
}

export interface IndustryCategory {
  industryCategoryId: number;

  industryCategoryName: string;

  isActive: boolean;
}